Unify toolbars into one class
