--- src/eterna/mode/FeedbackView/index.ts
@@ -1,2 +1,0 @@
-export {default as FeedbackViewMode} from "./FeedbackViewMode";
-export {default as FeedbackViewToolbar} from "./FeedbackViewToolbar"; "";
