+++ src/eterna/mode/PoseEdit/index.ts
@@ -2,7 +2,6 @@ export {default as Booster, BoosterType} from "./Booster";
 export {default as MissionClearedPanel} from "./MissionClearedPanel";
 export {default as MissionIntroMode} from "./MissionIntroMode";
 export {default as PoseEditMode, PuzzleState, PoseEditParams} from "./PoseEditMode";
-export {default as PoseEditToolbar} from "./PoseEditToolbar";
 export {default as SubmitPoseDetails} from "./SubmitPoseDetails";
 export {default as SubmitPoseDialog} from "./SubmitPoseDialog";
 export {default as SubmittingDialog} from "./SubmittingDialog";
