+++ src/eterna/mode/PuzzleEdit/index.ts
@@ -1,4 +1,3 @@
 export {default as PuzzleEditMode, PuzzleEditPoseData} from "./PuzzleEditMode";
-export {default as PuzzleEditToolbar} from "./PuzzleEditToolbar";
 export {default as StructureInput} from "./StructureInput";
 export {default as SubmitPuzzleDialog, SubmitPuzzleDetails} from "./SubmitPuzzleDialog";
