+++ src/eterna/mode/index.ts
@@ -3,6 +3,6 @@ export {default as ErrorDialogMode} from "./ErrorDialogMode";
 export {default as GameMode} from "./GameMode";
 export {default as LoadingMode} from "./LoadingMode";
 export {DesignBrowserMode, DesignBrowserFilter} from "./DesignBrowser";
-export {FeedbackViewMode} from "./FeedbackView";
+export {default as FeedbackViewMode} from "./FeedbackViewMode";
 export {PoseEditMode, PoseEditParams} from "./PoseEdit";
 export {PuzzleEditMode, PuzzleEditPoseData} from "./PuzzleEdit";
