+++ src/eterna/ui/EternaViewOptionsDialog.ts
@@ -34,9 +34,8 @@ export default class EternaViewOptionsDialog extends Dialog<void> {
         bind(Eterna.settings.highlightRestricted, "Highlight restricted sequences");
         bind(Eterna.settings.showChat, "In-game chat");
         bind(Eterna.settings.simpleGraphics, "Use simpler, less animated graphics");
-
+        bind(Eterna.settings.autohideToolbar, "Autohide toolbar");
         if (this._optionsMode !== EternaViewOptionsMode.PUZZLEMAKER) {
-            bind(Eterna.settings.autohideToolbar, "Autohide toolbar");
             bind(Eterna.settings.freezeButtonAlwaysVisible, "Freeze button always visible");
         }
